+++ src/App.tsx
@@ -3,6 +3,7 @@ import { Sidebar } from "@components/sidebar/sidebar";
 import { useXMTPClient } from "@hooks/use-xmtp-client";
 import { SidebarInset, SidebarProvider, useSidebar } from "@ui/sidebar";
 import { ConversationsProvider } from "@/src/contexts/xmtp-conversations-context";
+import { ToastProvider } from "@ui/toast";
 import { BrowserRouter, Routes, Route } from "react-router";
 import { ExplorePage } from "@components/explore/index";
 import { useSwipeGesture } from "@hooks/use-swipe-gesture";
@@ -98,9 +99,11 @@ export default function App() {
 
   return (
     <BrowserRouter>
-      <ConversationsProvider client={client}>
-        <AppContent />
-      </ConversationsProvider>
+      <ToastProvider>
+        <ConversationsProvider client={client}>
+          <AppContent />
+        </ConversationsProvider>
+      </ToastProvider>
     </BrowserRouter>
   );
 }